Added a method to get a recipe by ID.
Added a method to query a recipe from the db by recipe_id. Reworked the most_matches() method to call the get_recipe_by_id() method with the id of the recipe with the most matches with user's ingredients. The get_recipe_by_id() method then saves the recipe name, ingredients and instructions to the three instance variables which can be used to display in a GUI.